#379: put except_hook only on script or application startup and rename scripts with lbs prefix
