+++ docs/docs_config.py
@@ -37,10 +37,10 @@ import lbaf.IO.lbsVTDataReader as LoadReader
 import lbaf.IO.lbsVTDataWriter as VTDataWriter
 
 # Utilities
-import lbaf.Utils.exception_handler as exc_handler
+import lbaf.Utils.lbsExceptionHandler as exc_handler
 import lbaf.Utils.lbsCsv2JsonDataConverter as Csv2JsonConverter
 import lbaf.Utils.lbsDataStatFilesUpdater as DataStatFilesUpdater
-import lbaf.Utils.lbsLogger as logger
+import lbaf.Utils.lbsLogging as logger
 import lbaf.Utils.lbsVTDataExtractor as VTDataExtractor
