+++ src/lbaf/Execution/lbsInformAndTransferAlgorithm.py
@@ -1,11 +1,9 @@
-import sys
-
 from logging import Logger
+
+from ..IO.lbsStatistics import min_Hamming_distance, print_function_statistics
 from .lbsAlgorithmBase import AlgorithmBase
 from .lbsCriterionBase import CriterionBase
 from .lbsTransferStrategyBase import TransferStrategyBase
-from ..IO.lbsStatistics import print_function_statistics, min_Hamming_distance
-from ..Utils.exception_handler import exc_handler
 
 
 class InformAndTransferAlgorithm(AlgorithmBase):
@@ -34,18 +32,15 @@ class InformAndTransferAlgorithm(AlgorithmBase):
         if not isinstance(self.__n_iterations, int) or self.__n_iterations < 0:
             self._logger.error(
                 f"Incorrect provided number of algorithm iterations: {self.__n_iterations}")
-            sys.excepthook = exc_handler
             raise SystemExit(1)
         self.__n_rounds = parameters.get("n_rounds")
         if not isinstance(self.__n_rounds, int) or self.__n_rounds < 0:
             self._logger.error(
                 f"Incorrect provided number of information rounds: {self.__n_rounds}")
-            sys.excepthook = exc_handler
             raise SystemExit(1)
         self.__fanout = parameters.get("fanout")
         if not isinstance(self.__fanout, int) or self.__fanout < 0:
             self._logger.error(f"Incorrect provided information fanout {self.__fanout}")
-            sys.excepthook = exc_handler
             raise SystemExit(1)
         self._logger.info(
             f"Instantiated with {self.__n_iterations} iterations, {self.__n_rounds} rounds, fanout {self.__fanout}")
@@ -58,7 +53,6 @@ class InformAndTransferAlgorithm(AlgorithmBase):
             logger=self._logger)
         if not self.__transfer_criterion:
             self._logger.error(f"Could not instantiate a transfer criterion of type {crit_name}")
-            sys.excepthook = exc_handler
             raise SystemExit(1)
 
         # Try to instantiate object transfer strategy
@@ -70,7 +64,6 @@ class InformAndTransferAlgorithm(AlgorithmBase):
             lgr=self._logger)
         if not self.__transfer_strategy:
             self._logger.error(f"Could not instantiate a transfer strategy of type {strat_name}")
-            sys.excepthook = exc_handler
             raise SystemExit(1)
 
     def __information_stage(self):
