+++ src/lbaf/Execution/lbsPhaseStepperAlgorithm.py
@@ -1,9 +1,7 @@
-import sys
 from logging import Logger
 from .lbsAlgorithmBase import AlgorithmBase
 from ..Model.lbsPhase import Phase
 from ..IO.lbsStatistics import print_function_statistics
-from ..Utils.exception_handler import exc_handler
 
 
 class PhaseStepperAlgorithm(AlgorithmBase):
@@ -29,7 +27,6 @@ class PhaseStepperAlgorithm(AlgorithmBase):
                 [isinstance(p, Phase) for p in phases.values()]):
             self._logger.error(
                 "Algorithm execution requires a dictionary of phases")
-            sys.excepthook = exc_handler
             raise SystemExit(1)
 
         # Iterate over all phases
