+++ src/lbaf/Execution/lbsRecursiveTransferStrategy.py
@@ -1,16 +1,15 @@
-import sys
 import math
 import random
+from bisect import bisect
+from itertools import accumulate
 from logging import Logger
 from typing import Union
-from itertools import accumulate
-from bisect import bisect
 
-from .lbsTransferStrategyBase import TransferStrategyBase
-from ..Model.lbsPhase import Phase
-from ..Utils.exception_handler import exc_handler
 from ..IO.lbsStatistics import inverse_transform_sample
 from ..Model.lbsObjectCommunicator import ObjectCommunicator
+from ..Model.lbsPhase import Phase
+from .lbsTransferStrategyBase import TransferStrategyBase
+
 
 class RecursiveTransferStrategy(TransferStrategyBase):
     """A concrete class for the recursive transfer strategy."""
@@ -40,7 +39,6 @@ class RecursiveTransferStrategy(TransferStrategyBase):
         if o_s not in self.__strategy_mapped:
             self._logger.error(f"{o_s} does not exist in known ordering strategies: "
                                 f"{[x for x in self.__strategy_mapped]}")
-            sys.excepthook = exc_handler
             raise SystemExit(1)
         self.__order_strategy = self.__strategy_mapped[o_s]
         self._logger.info(f"Selected {self.__order_strategy.__name__} object ordering strategy")
