+++ src/lbaf/IO/lbsConfigurationValidator.py
@@ -1,10 +1,8 @@
 """LBAF Configuration validator."""
 from logging import Logger
-import sys
 from typing import Union, Dict, List
 
 from schema import And, Optional, Or, Regex, Schema, Use
-from ..Utils.exception_handler import exc_handler
 
 
 # Allowed configuration values
@@ -173,7 +171,6 @@ class ConfigurationValidator:
     @staticmethod
     def validate(valid_schema: Schema, schema_to_validate: dict):
         """Return validated schema."""
-        sys.excepthook = exc_handler
         return valid_schema.validate(schema_to_validate)
 
     @staticmethod
