+++ src/lbaf/Model/lbsWorkModelBase.py
@@ -1,9 +1,7 @@
 import abc
 from logging import Logger
-import sys
 
-from ..Utils.exception_handler import exc_handler
-from ..Utils.lbsLogger import get_logger
+from ..Utils.lbsLogging import get_logger
 
 
 class WorkModelBase:
@@ -23,10 +21,10 @@ class WorkModelBase:
     def factory(work_name, parameters, lgr: Logger):
         """Produce the necessary concrete work model."""
         # pylint:disable=W0641:possibly-unused-variable,C0415:import-outside-toplevel
-        from .lbsLoadOnlyWorkModel import LoadOnlyWorkModel
         from .lbsAffineCombinationWorkModel import AffineCombinationWorkModel
-        # pylint:enable=W0641:possibly-unused-variable,C0415:import-outside-toplevel
+        from .lbsLoadOnlyWorkModel import LoadOnlyWorkModel
 
+        # pylint:enable=W0641:possibly-unused-variable,C0415:import-outside-toplevel
         # Ensure that work name is valid
         try:
             # Instantiate and return object
@@ -35,7 +33,6 @@ class WorkModelBase:
         except Exception as err:
             # Otherwise, error out
             get_logger().error(f"Could not create a work with name: {work_name}")
-            sys.excepthook = exc_handler
             raise NameError(f"Could not create a work with name: {work_name}") from err
 
     @abc.abstractmethod
