+++ tests/unit/test_configuration_validator.py
@@ -6,7 +6,7 @@ import yaml
 
 from lbaf.Utils.path import abspath
 from lbaf.IO.lbsConfigurationValidator import ConfigurationValidator
-from lbaf.Utils.lbsLogger import get_logger
+from lbaf.Utils.lbsLogging import get_logger
 
 
 class TestConfig(unittest.TestCase):
